[Feature] New cvv screen (#2359)
* init feature branch

* security code fragment added

* viewmodel created

* paylogic added

* paylogic with paymentrecovery

* refactor ciew model factory

* revert change

* revert change in samples

* refactor

* review changes

* add resource string

* translation resource added

* test fixed

* retries with snackbar

* review changes

* snackbar error with action added

* review changes

* review changes

* SingleLiveData created

* MutableSingleLiveData created

* add tab

* [Feature] payment logic (#2313)

* init feature branch

* security code fragment added

* viewmodel created

* paylogic added

* paylogic with paymentrecovery

* refactor ciew model factory

* revert change

* revert change in samples

* refactor

* review changes

* add resource string

* translation resource added

* test fixed

* review changes

* review changes

* review changes

* SingleLiveData created

* MutableSingleLiveData created

* add tab

* refactor code

* revert changes

* test fixed

* refactor code

* handle back fixed

* refactor

* review change

* Usecase added

* refactor usecase

* tracks

* error handler fixed

* revert change

* clear cap added

* test fixed

* refactor tracks

* refactor ConnectionError

* remove values from paybutton

* [Feature] Retries and error handling (#2321)

* init feature branch

* security code fragment added

* viewmodel created

* paylogic added

* paylogic with paymentrecovery

* refactor ciew model factory

* revert change

* revert change in samples

* refactor

* review changes

* add resource string

* translation resource added

* test fixed

* retries with snackbar

* review changes

* snackbar error with action added

* review changes

* review changes

* SingleLiveData created

* MutableSingleLiveData created

* add tab

* refactor code

* revert changes

* test fixed

* refactor code

* handle back fixed

* refactor

* review change

* error handler fixed

* revert change

* clear cap added

* test fixed

* refactor ConnectionError

* remove values from paybutton

* refactor

* change name method in usecase

* change package Response class

* [Feature] Use cases (#2335)

* init feature branch

* security code fragment added

* viewmodel created

* paylogic added

* paylogic with paymentrecovery

* refactor ciew model factory

* revert change

* revert change in samples

* refactor

* review changes

* add resource string

* translation resource added

* test fixed

* retries with snackbar

* review changes

* snackbar error with action added

* review changes

* review changes

* SingleLiveData created

* MutableSingleLiveData created

* add tab

* refactor code

* revert changes

* test fixed

* refactor code

* handle back fixed

* refactor

* review change

* Usecase added

* refactor usecase

* error handler fixed

* revert change

* clear cap added

* test fixed

* refactor ConnectionError

* remove values from paybutton

* change name method in usecase

* change package Response class

* refactor

* review changes

* review code

* Added security code fragment transition

* Added correct subtitle + groups support

* test for tokenize use case

* Cleaned animation code

* use case test added

* Fixed tests

* Disable and enable button logic in cvv screen

* viewmodel, tracking and mappers test are added

* Refactor animator factory into extensions

* [Feature] Translates (#2350)

* update translations

* refactor en translations

* refactor

* Added new SecurityCodeActivity for pay button integrations without fragment container

* Fix front location cvv on one tap and security code fragments

* onetap animation trigger refactor

* review chnages

* update notNull function

* [Feature] Override state pay button (#2362)

* override state pay button

* revert changes

* review changes

* refactor

* change name

* [Feature] Update andes version (#2363)

* Update andes version

* refactor

* update test

+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/view/SummaryView.java
@@ -116,7 +116,7 @@ public class SummaryView extends LinearLayout {
         toolbarElementDescriptor.setVisibility(VISIBLE);
     }
 
-    public void animateEnter(final int duration) {
+    public void animateEnter(final long duration) {
         shouldAnimateReturnFromCardForm = true;
         detailAdapter.customAnimation = true;
 
@@ -146,7 +146,7 @@ public class SummaryView extends LinearLayout {
         totalAmountDescriptor.animateEnter();
     }
 
-    public void animateExit(final int duration) {
+    public void animateExit(final long duration) {
         final Animation fadeOut = AnimationUtils.loadAnimation(getContext(), R.anim.px_fade_out);
         fadeOut.setDuration(duration);
